moved constants to config.py, SerialDeviceConnection class to serial_connection.py and ScalableGroup(and globals params + p) to params.py. Changed nothing functionally
