+++ src/app.py
@@ -22,315 +22,10 @@ from PyQt5.QtWidgets import (QMainWindow, QSplitter, QApplication, QTabWidget, Q
 from pyqtgraph import GraphicsLayoutWidget, DateAxisItem, AxisItem, ViewBox, PlotCurveItem, LegendItem, PlotItem, mkPen, mkBrush
 from pyqtgraph.parametertree import Parameter, ParameterTree, parameterTypes
 
-# current version number displayed in the GUI (Major.Minor.Patch or Breaking.Feature.Fix)
-version_number = "0.10.9"
+from config import *
+from serial_connection import SerialDeviceConnection
+from params import ScalableGroup, params, p
 
-# Define instrument types
-CPC = 1
-PSM = 2
-Electrometer = 3
-CO2_sensor = 4
-RHTP = 5
-eDiluter = 6
-PSM2 = 7
-TSI_CPC = 8
-AFM = 9
-Example_device = -1
-
-# self test error descriptions
-CPC_ERRORS = (
-    "RESERVED FOR FUTURE USE", "ERROR_SELFTEST_FLASH_ID", "ERROR_SELFTEST_TEMP_OPTICS", "ERROR_SELFTEST_TEMP_SATURATOR", "ERROR_SELFTEST_TEMP_CONDENSER",
-    "ERROR_SELFTEST_TEMP_BOARD", "ERROR_SELFTEST_LIQUID_SENSOR", "ERROR_SELFTEST_PRESSURE_INLET", "ERROR_SELFTEST_PRESSURE_NOZZLE", "ERROR_SELFTEST_PRESSURE_CRITICAL",
-    "ERROR_SELFTEST_DISPLAY", "ERROR_SELFTEST_VOLTAGE_3V3", "ERROR_SELFTEST_VOLTAGE_5V", "ERROR_SELFTEST_VOLTAGE_12V", "ERROR_SELFTEST_VOLTAGE_REF_NTC",
-    "ERROR_SELFTEST_VOLTAGE_REF_PRES", "ERROR_SELFTEST_VOLTAGE_REF_DAC", "ERROR_SELFTEST_VOLTAGE_OPC_DC", "ERROR_SELFTEST_VOLTAGE_LASER",
-    "ERROR_SELFTEST_FAN1", "ERROR_SELFTEST_FAN2", "ERROR_SELFTEST_FAN3", "ERROR_SELFTEST_PRESSURE_CAB", "ERROR_SELFTEST_RTC",
-    "ERROR_SELFTEST_DAC1", "ERROR_SELFTEST_DAC2", "ERROR_SELFTEST_ANALOG_OUT", "ERROR_SELFTEST_ANALOG_OUT2", "ERROR_SELFTEST_PRESSURE_WREM",
-)
-
-PSM_ERRORS = (
-    "RESERVED FOR FUTURE USE", "ERROR_SELFTEST_FLASH_ID", "ERROR_SELFTEST_TEMP_GROWTH_TUBE", "ERROR_SELFTEST_TEMP_SATURATOR", "RESERVED FOR FUTURE USE",
-    "ERROR_SELFTEST_TEMP_BOARD", "ERROR_SELFTEST_LIQUID_SENSOR", "ERROR_SELFTEST_PRESSURE_ABS", "ERROR_SELFTEST_PRESSURE_ABSSAT", "ERROR_SELFTEST_PRESSURE_SATDRY",
-    "ERROR_SELFTEST_TEMP_PREHEATER", "ERROR_SELFTEST_VOLTAGE_3V3", "ERROR_SELFTEST_VOLTAGE_5V", "ERROR_SELFTEST_VOLTAGE_12V", "ERROR_SELFTEST_VOLTAGE_REF_NTC",
-    "ERROR_SELFTEST_VOLTAGE_REF_PRES", "ERROR_SELFTEST_VOLTAGE_REF_DAC", "ERROR_SELFTEST_TEMP_INLET", "ERROR_SELFTEST_DRAIN_LIQUID_SENSOR",
-    "ERROR_SELFTEST_FAN1", "ERROR_SELFTEST_FAN2", "ERROR_SELFTEST_FAN3", "ERROR_SELFTEST_RTC", "ERROR_SELFTEST_DAC1", "ERROR_SELFTEST_DAC2",
-    "ERROR_SELFTEST_TEMP_DRAIN", "ERROR_SELFTEST_MFC_SATURATOR", "ERROR_SELFTEST_MFC_HEATER", "ERROR_SELFTEST_PRESSURE_CRIT", "ERROR_SELFTEST_MFC_VACUUM"
-)
-
-warnings.filterwarnings("ignore", message='Mean of empty slice')
-warnings.filterwarnings("ignore", message='All-NaN slice encountered')
-
-# Set the LC_ALL environment variable to US English (en_US)
-locale.setlocale(locale.LC_ALL, 'en_US.UTF-8')
-
-# set up logging
-logging.basicConfig(filename='debug.log', encoding='UTF-8', level=logging.DEBUG, format='%(asctime)s %(levelname)s %(message)s', datefmt='%m/%d/%Y %I:%M:%S %p')
-
-# define file paths according to run mode (exe or script)
-script_path = os.path.realpath(os.path.dirname(__file__)) # location of this file
-# exe (one file)
-if getattr(sys, 'frozen', False):
-    save_path = os.path.realpath(os.path.dirname(sys.executable)) # save files to exe location
-    resource_path = script_path + "/res" # path of /res/ folder (images, icons)
-# script
-else:
-    save_path = os.path.dirname(script_path) # save files to repository's main folder
-    resource_path = os.path.dirname(script_path) + "/res" # path of /res/ folder (images, icons)
-
-# check if platform is OSX
-if platform.system() == "Darwin":
-    # OSX mode makes code compatible with OSX
-    osx_mode = 1
-else:
-    osx_mode = 0
-
-class SerialDeviceConnection():
-    def __init__(self):
-        self.serial_port = "NaN"
-        self.timeout = 0.2
-        self.baud_rate = 115200
-        
-    def set_port(self, serial_port):
-        self.serial_port = serial_port
-    
-    def set_baud_rate(self, baud_rate):
-        self.baud_rate = baud_rate
-    
-    # open serial connection
-    def connect(self):
-        try: # Try to close with the port that was last used (needed if the port has been changed)
-            self.connection.close()
-        except: # if fails (i.e. port has not been open) continue normally
-            pass
-        self.connection = Serial(self.serial_port, self.baud_rate, timeout=self.timeout) #, rtscts=True)
-        print("Connected to %s" % self.serial_port)
-    
-    # close serial connection
-    def close(self):
-        # if connection exist, it is closed
-        try:
-            # Try to close with the port that was last used (needed if the port has been changed)
-            self.connection.close()
-            #print("Connection closed")
-        except:
-            pass
-    
-    # close old connection and open new connection
-    def change_port(self, serial_port):
-        # close current connection
-        self.close()
-        # set new port
-        self.set_port(serial_port)
-        try:
-            # connect to new port
-            self.connect()
-        except Exception as e:
-            print("change_port:", e)
-    
-    def send_message(self, message):
-        # add line termination and convert to bytes
-        message = bytes((str(message)+'\r\n'), 'utf-8')
-        try:
-            # send message if connection exists
-            self.connection.write(message)
-        except AttributeError:
-            # print message if connection does not exist
-            print("send_message - no connection, message -", message)
-    
-    def send_delayed_message(self, message, delay):
-        QTimer.singleShot(delay, lambda: self.send_message(message))
-    
-    def send_multiple_messages(self, device_type, ten_hz=False):
-
-        if device_type == 1: # CPC
-            self.send_message(":MEAS:ALL")
-            QTimer.singleShot(150, lambda: self.send_message(":SYST:PRNT"))
-            QTimer.singleShot(300, lambda: self.send_message(":SYST:PALL"))
-            if ten_hz:
-                QTimer.singleShot(450, lambda: self.send_message(":MEAS:OPC_CONC_LOG"))
-        
-        elif device_type == TSI_CPC: # TSI CPC
-            self.send_message("RD") # read concentration
-            QTimer.singleShot(150, lambda: self.send_message("RIE")) # read instrument errors
-    
-    def send_pulse_analysis_messages(self, threshold):
-        # send required messages for pulse analysis
-        self.send_message(":SET:OPC:THRS " + str(threshold))
-        QTimer.singleShot(150, lambda: self.send_message(":MEAS:ALL"))
-    
-    # --- CPC & PSM set/command functions ---
-
-    # # send set message
-    def send_set(self, message):
-        if message == None:
-            # if message is None, do nothing
-            return
-        else:
-            # send message
-            self.send_message(message)
-    
-    # add value to set message
-    def send_set_val(self, value, message, **kwargs):
-        if isinstance(value, float): # if value is float, round to 2 or x decimals
-            if kwargs: # if kwargs is not empty
-                value = round(value, kwargs['decimals']) # round value to kwargs['decimals'] decimals
-            else: # otherwise, round to 2 decimals
-                value = round(value, 2)
-        message = message + str(value) # add value to message
-        #print("send_set_val -", message) # print message for debugging
-        self.send_set(message) # send message using send_message()
-
-# ScalableGroup for creating a menu where to set up new COM devices
-class ScalableGroup(parameterTypes.GroupParameter):
-    def __init__(self, **opts):
-        #opts['type'] = 'action'
-        opts['addText'] = "Add new device"
-        # opts for choosing device type when adding new device
-        opts["addList"] = ["CPC", "PSM Retrofit", "PSM 2.0", "Electrometer", "CO2 sensor", "RHTP", "AFM", "eDiluter", "TSI CPC", "Example device"]
-        parameterTypes.GroupParameter.__init__(self, **opts)
-        self.n_devices = 0
-        self.cpc_dict = {'None': 'None'}
-        # update cpc_dict when device is removed
-        self.sigChildRemoved.connect(self.update_cpc_dict)
-
-    def addNew(self, device_type, device_name=None): # device_type is the name of the added device type
-        # device_value is used to set the default value for the Device type parameter below
-        device_value = {"CPC": CPC, "PSM Retrofit": PSM, "PSM 2.0": PSM2, "Electrometer": Electrometer, "CO2 sensor": CO2_sensor, "RHTP": RHTP, "AFM": AFM, "eDiluter": eDiluter, "TSI CPC": TSI_CPC, "Example device": -1}[device_type]
-        # if OSX mode is on, set COM port type as string to allow complex port addresses
-        if osx_mode:
-            port_type = 'str'
-        else:
-            port_type = 'int'
-        # if device_name argument is not given, set device name according to device type
-        if device_name == None:
-            device_name = device_type
-        # if device name is in use, add number to the end of the name
-        if device_name in [child.name() for child in self.children()]:
-            name_number = 1
-            name_set = False
-            while not name_set:
-                name_number += 1
-                if device_name + " (%d)" % (name_number) not in [child.name() for child in self.children()]:
-                    device_name = device_name + " (%d)" % (name_number)
-                    name_set = True
-        # New types of devices should be added in the "Device type" list and given unique id number
-        self.addChild({'name': device_name, 'removable': True, 'type': 'group', 'children': [
-                dict(name="Device nickname", type='str', value="", renamable=True),
-                dict(name="COM port", type=port_type),
-                dict(name="Serial number", type='str', value="", readonly=True),
-                #dict(name="Baud rate", type='int', value=115200, visible=False),
-                dict(name = "Connection", value = SerialDeviceConnection(), visible=False),
-                {'name': 'Device type', 'type': 'list', 'values': {"CPC": CPC, "PSM Retrofit": PSM, "PSM 2.0": PSM2, "Electrometer": Electrometer, "CO2 sensor": CO2_sensor, "RHTP": RHTP, "AFM": AFM, "eDiluter": eDiluter, "TSI CPC": TSI_CPC, "Example device": -1}, 'value': device_value, 'readonly': True, 'visible': False},
-                dict(name = "Connected", type='bool', value=False, readonly = True),
-                dict(name = "DevID", type='int', value=self.n_devices,readonly = True, visible = False),
-                dict(name = "Plot to main", type='bool', value=True),
-                ]})
-
-        self.n_devices += 1 # increase device counter
-
-        # if added device is CPC, update cpc_dict
-        if device_value in [CPC, TSI_CPC]:
-            self.update_cpc_dict()
-            # if Airmodus CPC, add hidden 10 hz parameter
-            # when 10 hz is True, OPC concentration is polled
-            if device_value == CPC:
-                self.children()[-1].addChild({'name': '10 hz', 'type': 'bool', 'value': False, 'readonly': True, 'visible': False})
-
-        # if added device is PSM, add hidden parameters and option for 'Connected CPC'
-        if device_value in [PSM, PSM2]:
-            # if device is PSM Retrofit, add hidden CO flow parameter
-            if device_value == PSM:
-                self.children()[-1].addChild({'name': 'CO flow', 'type': 'str', 'visible': False})
-            # add hidden 10 hz parameter for storing 10 hz status for startup
-            self.children()[-1].addChild({'name': '10 hz', 'type': 'bool', 'value': False, 'readonly': True, 'visible': False})
-            # add options for connected CPC
-            self.children()[-1].addChild({'name': 'Connected CPC', 'type': 'list', 'values': self.cpc_dict, 'value': 'None'})
-            # add cpc_changed flag to device
-            self.children()[-1].cpc_changed = False
-            # connect value change signal of Connected CPC to update_cpc_changed slot
-            self.children()[-1].child('Connected CPC').sigValueChanged.connect(self.update_cpc_changed)
-            # add firmware version parameter to index 3
-            self.children()[-1].insertChild(3, {'name': 'Firmware version', 'type': 'str', 'value': "", 'readonly': True})
-        
-        # if added device is RHTP, add options for plotted value
-        if device_value == RHTP:
-            # remove default Plot to main parameter
-            self.children()[-1].removeChild(self.children()[-1].child('Plot to main'))
-            # create new Plot to main parameter with options for plotted value
-            self.children()[-1].addChild({'name': 'Plot to main', 'type': 'list', 'values': [None, 'RH', 'T', 'P'], 'value': 'RH'})
-        
-        # if added device is AFM, add options for plotted value
-        if device_value == AFM:
-            # remove default Plot to main parameter
-            self.children()[-1].removeChild(self.children()[-1].child('Plot to main'))
-            # create new Plot to main parameter with options for plotted value
-            self.children()[-1].addChild({'name': 'Plot to main', 'type': 'list', 'values': [None, 'Flow', 'Standard flow', 'RH', 'T', 'P'], 'value': 'Flow'})
-        
-        # if added device is Example device, hide irrelevant parameters
-        if device_value == Example_device:
-            self.children()[-1].child('COM port').setOpts(visible=False)
-            self.children()[-1].child('Serial number').setOpts(visible=False)
-            self.children()[-1].child('Connected').setOpts(visible=False)
-
-    def update_cpc_dict(self):
-        self.cpc_dict = {'None': 'None'} # reset cpc_dict
-        # add device name to cpc_dict if device is CPC
-        for device in self.children():
-            if device.child('Device type').value() in [CPC, TSI_CPC]:
-                # check if device has a nickname
-                if device.child('Device nickname').value() != "":
-                    self.cpc_dict[device.child('Device nickname').value()] = device.child('DevID').value()
-                else: # if no nickname, use device parameter name (device type and serial number)
-                    self.cpc_dict[device.name()] = device.child('DevID').value()
-        # update Connected CPC parameter for all PSM devices
-        for device in self.children():
-            if device.child('Device type').value() in [PSM, PSM2]:
-                # store current value (ID) of Connected CPC parameter
-                current_cpc = device.child('Connected CPC').value()
-                # remove Connected CPC parameter
-                device.removeChild(device.child('Connected CPC'))
-                # add updated Connected CPC parameter
-                device.addChild({'name': 'Connected CPC', 'type': 'list', 'values': self.cpc_dict})
-                # set Connected CPC parameter to previous value if it is still in the list
-                if current_cpc in self.cpc_dict.values():
-                    device.child('Connected CPC').setValue(current_cpc)
-                else: # else set cpc_changed to True
-                    device.cpc_changed = True
-                # connect value change signal of Connected CPC to update_cpc_changed slot
-                device.child('Connected CPC').sigValueChanged.connect(self.update_cpc_changed)
-
-    # slot for setting cpc_changed flag to True when Connected CPC parameter is changed
-    def update_cpc_changed(self, value):
-        device = value.parent() # get device parameter
-        device.cpc_changed = True # set cpc_changed flag to True
-
-# Create a dictionary, in which the names, types and default values are set
-params = [
-    {'name': 'Data settings', 'type': 'group', 'children': [
-        {'name': 'File path', 'type': 'str', 'value': save_path},
-        {'name': 'File tag', 'type': 'str', 'value': "", 'tip': "File name: YYYYMMDD_HHMMSS_(Serial number)_(Device type)_(Device nickname)_(File tag).dat"},
-        {'name': 'Save data', 'type': 'bool', 'value': False},
-        {'name': 'Generate daily files', 'type': 'bool', 'value': True, 'tip': "If on, new files are started at midnight."},
-        {'name': 'Resume on startup', 'type': 'bool', 'value': False, 'tip': "Option to resume the last settings on startup."},
-        {'name': 'Save settings', 'type': 'action'},
-        {'name': 'Load settings', 'type': 'action'},
-    ]},
-    {'name': 'Plot settings', 'type': 'group', 'children': [
-        {'name': 'Follow', 'type': 'bool', 'value': True},
-        {'name': 'Time window (s)', 'type': 'int', 'value': 60},
-        {'name': 'Autoscale Y', 'type': 'bool', 'value': True}
-    ]},
-    {'name': 'Serial ports', 'type': 'group', 'children': [
-        {'name': 'Available serial ports', 'type': 'text', 'value': '', 'readonly': True},
-        {'name': 'Update serial ports', 'type': 'action'},
-    ]},
-    
-    ScalableGroup(name="Device settings", children=[
-        # devices will be added here
-    ]),
-]
-
-# Create tree of Parameter objects
-p = Parameter.create(name='params', type='group', children=params)
 
 # main program
 class MainWindow(QMainWindow):
